requirements.txt
alabaster==0.7.10
altgraph==0.16.1
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==1.6.3
astropy==3.0.2
atomicwrites==1.3.0
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
ConcurrentLogHandler==0.9.1
conda==4.7.12
conda-build==3.10.5
conda-package-handling==1.6.0
conda-verify==2.0.0
configparser==3.7.4
contextlib2==0.5.5
cryptography==2.8
cx-Freeze==6.1
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
distributed==1.21.8
Django==3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
docopt==0.6.2
docutils==0.14
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.4
fire==0.1.3
Flask==1.0.2
Flask-Cors==3.0.4
future==0.17.1
gevent==1.3.0
glob2==0.6
greenlet==0.4.13
h5py==2.7.1
hdfs==2.5.2
heapdict==1.0.0
html5lib==1.0.1
HTMLParser==0.0.2
idna==2.8
imageio==2.3.0
imagesize==1.0.0
importlib-metadata==0.23
ipykernel==4.8.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.13.3
jieba==0.39
Jinja2==2.10
joblib==0.13.2
json5==0.8.5
jsonpath==0.82
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
jupyterlab-server==1.0.0
kazoo==2.5.0
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
mysqlclient==1.4.6
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.7.8
numba==0.38.0
numexpr==2.7.0
numpy==1.16.4
numpydoc==0.8.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pefile==2019.4.18
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.4.5
ptyprocess==0.6.0
pure-sasl==0.6.1
py==1.5.3
py2exe==0.9.2.2
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==1.6.0
Pygments==2.4.2
PyHive==0.6.1
PyInstaller==3.6
pykafka==2.8.0
pylint==1.8.4
pymongo==3.8.0
PyMySQL==0.9.3
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==0.5.2
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
redis==2.10.6
requests==2.22.0
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
soupsieve==1.9.3
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
spyder-kernels==1.4.0
SQLAlchemy==1.2.7
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.3
TBB==0.1
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
thrift==0.11.0
thrift-sasl==0.3.0
toolz==0.9.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
typed-ast==1.4.0
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wfastcgi==3.0.0
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zipp==0.6.0
